refactor(web): authenticateのPublic APIを単一バレルに統一
/ui サブバレルへの直接importをやめ、consumer は @/client/features/authenticate
のみを使う。バレル経由で storybook バンドルに巻き込まれる非UI資産（feature直下の
hook・検証・エラーメッセージ・turnstile と client/infrastructure）は client
ユニットテストで担保するため storybook の coverage 母数から除外し、CJS interop
回避のため optimizeDeps に swr / pino を追加する。
