+++ application/packages/web/src/client/features/authenticate/index.ts
@@ -1,3 +1,5 @@
+export { AuthenticateForm } from './components/authenticate-form'
+export { default as LogoutButton } from './components/logout-button'
 export {
   AUTH_ERROR_MESSAGES,
   resolveLoginErrorMessage,
