--- application/packages/web/src/client/features/authenticate/ui.ts
@@ -1,2 +1,0 @@
-export { AuthenticateForm } from './components/authenticate-form'
-export { default as LogoutButton } from './components/logout-button'
